Fix linting errors and improve type safety
- Fixed react-refresh warnings by moving metadata exports to separate files
- Replaced all 'any' types with proper TypeScript types in PerformanceMonitor and monitoring utils
- Improved type safety for Performance API entries
- All linting errors resolved, tests passing, type check clean
